+++ client/client.go
@@ -16,16 +16,22 @@ import (
 )
 
 import (
+	"github.com/AlexStocks/dubbogo/codec"
+	"github.com/AlexStocks/dubbogo/codec/hessian"
+	"github.com/AlexStocks/dubbogo/codec/jsonrpc"
 	"github.com/AlexStocks/dubbogo/registry"
+	"github.com/AlexStocks/dubbogo/registry/zk"
+	"github.com/AlexStocks/dubbogo/selector"
+	"github.com/AlexStocks/dubbogo/selector/cache"
+	"github.com/AlexStocks/dubbogo/transport"
 )
 
 // Client is the interface used to make requests to services.
 // It supports Request/Response via Transport and Publishing via the Broker.
 // It also supports bidirectional streaming of requests.
 type Client interface {
-	Init(...Option) error
 	Options() Options
-	NewJsonRequest(service, method string, req interface{}, reqOpts ...RequestOption) Request
+	NewRequest(service, method string, req interface{}, reqOpts ...RequestOption) Request
 	Call(ctx context.Context, req Request, rsp interface{}, opts ...CallOption) error
 	String() string
 	Close()
@@ -52,14 +58,23 @@ type Streamer interface {
 	Close() error
 }
 
-// Option used by the Client
-type Option func(*Options)
+type (
+	// Option used by the Client
+	Option func(*Options)
+	// CallOption used by Call or Stream
+	CallOption func(*CallOptions)
+	// RequestOption used by NewRequest
+	RequestOption func(*RequestOptions)
+)
 
-// CallOption used by Call or Stream
-type CallOption func(*CallOptions)
-
-// RequestOption used by NewRequest
-type RequestOption func(*RequestOptions)
+type (
+	dubbogoClientConfig struct {
+		codecType     codec.CodecType
+		newCodec      codec.NewCodec
+		transportType codec.TransportType // transport type
+		newTransport  transport.NewTransport
+	}
+)
 
 var (
 	// DefaultRetries is the default number of times a request is tried
@@ -70,6 +85,41 @@ var (
 	DefaultPoolSize = 0
 	// DefaultPoolTTL sets the connection pool ttl
 	DefaultPoolTTL = time.Minute
+
+	contentType2Codec = map[string]codec.NewCodec{
+		"application/json":    jsonrpc.NewCodec,
+		"application/jsonrpc": jsonrpc.NewCodec,
+		"application/dubbo":   hessian.NewCodec,
+	}
+
+	codec2ContentType = map[string]string{
+		"jsonrpc": "application/json",
+		"dubbo":   "application/dubbo",
+	}
+
+	dubbogoClientConfigMap = map[codec.CodecType]dubbogoClientConfig{
+		codec.CODECTYPE_JSONRPC: dubbogoClientConfig{
+			codecType:     codec.CODECTYPE_JSONRPC,
+			newCodec:      jsonrpc.NewCodec,
+			transportType: codec.TRANSPORT_HTTP,
+			newTransport:  transport.NewHTTPTransport,
+		},
+
+		codec.CODECTYPE_DUBBO: dubbogoClientConfig{
+			codecType:     codec.CODECTYPE_DUBBO,
+			newCodec:      hessian.NewCodec,
+			transportType: codec.TRANSPORT_TCP,
+			newTransport:  transport.NewTcpTransport,
+		},
+	}
+
+	DefaultRegistries = map[string]registry.NewRegistry{
+		"zookeeper": zookeeper.NewConsumerZookeeperRegistry,
+	}
+
+	DefaultSelectors = map[string]selector.NewSelector{
+		"cache": cache.NewSelector,
+	}
 )
 
 // creates a new client with the options passed in