+++ server/server.go
@@ -60,7 +60,7 @@ func newServer(opts ...Option) Server {
 	}
 }
 
-func (this *server) handlePkg(servo interface{}, sock transport.Socket) {
+func (s *server) handlePkg(servo interface{}, sock transport.Socket) {
 	var (
 		ok          bool
 		rpc         *rpcServer
@@ -98,11 +98,11 @@ func (this *server) handlePkg(servo interface{}, sock transport.Socket) {
 		}
 
 		// 下面的所有逻辑都是处理请求包，并回复response
-		// we use this Content-Type header to identify the codec needed
+		// we use s Content-Type header to identify the codec needed
 		contentType = msg.Header["Content-Type"]
 
 		// codec of jsonrpc & other type etc
-		codecFunc, err = this.newCodec(contentType)
+		codecFunc, err = s.newCodec(contentType)
 		if err != nil {
 			sock.Send(&transport.Message{
 				Header: map[string]string{
@@ -127,7 +127,7 @@ func (this *server) handlePkg(servo interface{}, sock transport.Socket) {
 
 		// ctx = metadata.NewContext(context.Background(), header)
 		ctx = context.WithValue(context.Background(), common.DUBBOGO_CTX_KEY, header)
-		// we use this Timeout header to set a server deadline
+		// we use s Timeout header to set a server deadline
 		if len(msg.Header["Timeout"]) > 0 {
 			if timeout, err = strconv.ParseUint(msg.Header["Timeout"], 10, 64); err == nil {
 				ctx, _ = context.WithTimeout(ctx, time.Duration(timeout))
@@ -141,12 +141,12 @@ func (this *server) handlePkg(servo interface{}, sock transport.Socket) {
 	}
 }
 
-func (this *server) newCodec(contentType string) (codec.NewCodec, error) {
+func (s *server) newCodec(contentType string) (codec.NewCodec, error) {
 	var (
 		ok bool
 		cf codec.NewCodec
 	)
-	if cf, ok = this.opts.Codecs[contentType]; ok {
+	if cf, ok = s.opts.Codecs[contentType]; ok {
 		return cf, nil
 	}
 	if cf, ok = defaultCodecs[contentType]; ok {
@@ -155,22 +155,22 @@ func (this *server) newCodec(contentType string) (codec.NewCodec, error) {
 	return nil, jerrors.Errorf("Unsupported Content-Type: %s", contentType)
 }
 
-func (this *server) Options() Options {
+func (s *server) Options() Options {
 	var (
 		opts Options
 	)
 
-	this.RLock()
-	opts = this.opts
-	this.RUnlock()
+	s.RLock()
+	opts = s.opts
+	s.RUnlock()
 
 	return opts
 }
 
 /*
 type ProviderServiceConfig struct {
-	Protocol string // from ServiceConfig, get field{Path} from ServerConfig by this field
-	Service string  // from handler, get field{Protocol, Group, Version} from ServiceConfig by this field
+	Protocol string // from ServiceConfig, get field{Path} from ServerConfig by s field
+	Service string  // from handler, get field{Protocol, Group, Version} from ServiceConfig by s field
 	Group   string
 	Version string
 	Methods string
@@ -191,7 +191,7 @@ type ServerConfig struct {
 }
 */
 
-func (this *server) Handle(h Handler) error {
+func (s *server) Handle(h Handler) error {
 	var (
 		i           int
 		j           int
@@ -202,7 +202,7 @@ func (this *server) Handle(h Handler) error {
 		config      Options
 		serviceConf registry.ProviderServiceConfig
 	)
-	config = this.Options()
+	config = s.Options()
 
 	serviceConf.Service = h.Service()
 	serviceConf.Version = h.Version()
@@ -219,9 +219,9 @@ func (this *server) Handle(h Handler) error {
 			// serviceConf.Version = config.ServiceConfList[i].Version
 			for j = 0; j < serverNum; j++ {
 				if config.ServerConfList[j].Protocol == serviceConf.Protocol {
-					this.Lock()
-					serviceConf.Methods, err = this.rpc[j].register(h)
-					this.Unlock()
+					s.Lock()
+					serviceConf.Methods, err = s.rpc[j].register(h)
+					s.Unlock()
 					if err != nil {
 						return err
 					}
@@ -241,14 +241,14 @@ func (this *server) Handle(h Handler) error {
 		return jerrors.Errorf("fail to register Handler{service:%s, version:%s}", serviceConf.Service, serviceConf.Version)
 	}
 
-	this.Lock()
-	this.handlers[h.Service()] = h
-	this.Unlock()
+	s.Lock()
+	s.handlers[h.Service()] = h
+	s.Unlock()
 
 	return nil
 }
 
-func (this *server) Start() error {
+func (s *server) Start() error {
 	var (
 		i         int
 		serverNum int
@@ -257,7 +257,7 @@ func (this *server) Start() error {
 		rpc       *rpcServer
 		listener  transport.Listener
 	)
-	config = this.Options()
+	config = s.Options()
 
 	serverNum = len(config.ServerConfList)
 	for i = 0; i < serverNum; i++ {
@@ -267,43 +267,43 @@ func (this *server) Start() error {
 		}
 		log.Info("Listening on %s", listener.Addr())
 
-		this.Lock()
-		rpc = this.rpc[i]
+		s.Lock()
+		rpc = s.rpc[i]
 		rpc.listener = listener
-		this.Unlock()
+		s.Unlock()
 
-		this.wg.Add(1)
+		s.wg.Add(1)
 		go func(servo *rpcServer) {
-			listener.Accept(func(s transport.Socket) { this.handlePkg(rpc, s) })
-			this.wg.Done()
+			listener.Accept(func(sock transport.Socket) { s.handlePkg(rpc, sock) })
+			s.wg.Done()
 		}(rpc)
 
-		this.wg.Add(1)
+		s.wg.Add(1)
 		go func(servo *rpcServer) { // server done goroutine
 			var err error
-			<-this.done                  // step1: block to wait for done channel(wait server.Stop step2)
+			<-s.done                     // step1: block to wait for done channel(wait server.Stop step2)
 			err = servo.listener.Close() // step2: and then close listener
 			if err != nil {
 				log.Warn("listener{addr:%s}.Close() = error{%#v}", servo.listener.Addr(), err)
 			}
-			this.wg.Done()
+			s.wg.Done()
 		}(rpc)
 	}
 
 	return nil
 }
 
-func (this *server) Stop() {
-	this.once.Do(func() {
-		close(this.done)
-		this.wg.Wait()
-		if this.opts.Registry != nil {
-			this.opts.Registry.Close()
-			this.opts.Registry = nil
+func (s *server) Stop() {
+	s.once.Do(func() {
+		close(s.done)
+		s.wg.Wait()
+		if s.opts.Registry != nil {
+			s.opts.Registry.Close()
+			s.opts.Registry = nil
 		}
 	})
 }
 
-func (this *server) String() string {
+func (s *server) String() string {
 	return "dubbogo rpc server"
 }