+++ change_log.md
@@ -16,6 +16,7 @@ dubbogo 目前版本(0.1.1)支持的codec 是jsonrpc 2.0，transport protocol是
 - 2 delete ListService
 - 3 Registry.GetService -> Registry.GetServices
 - 4 use ServiceConfigIf in selector.Select()
+- 5 把dubbog-examples client下面一些默认配置放到client package 下面
 
 ### 2016-10-26 ###
 ---